# openapi/openapi.snapshot.json
{
  "components": {
    "schemas": {
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "IncidentEvidenceNew": {
        "properties": {
          "counts": {
            "additionalProperties": {
              "type": "integer"
            },
            "title": "Counts",
            "type": "object"
          },
          "events": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Events",
            "type": "array"
          },
          "timeline": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Timeline",
            "type": "array"
          },
          "window_end": {
            "title": "Window End",
            "type": "string"
          },
          "window_start": {
            "title": "Window Start",
            "type": "string"
          }
        },
        "required": [
          "window_start",
          "window_end",
          "counts",
          "timeline",
          "events"
        ],
        "title": "IncidentEvidenceNew",
        "type": "object"
      },
      "IncidentExplanationNew": {
        "properties": {
          "observed": {
            "title": "Observed",
            "type": "integer"
          },
          "threshold": {
            "title": "Threshold",
            "type": "integer"
          },
          "trigger_field": {
            "title": "Trigger Field",
            "type": "string"
          },
          "window": {
            "title": "Window",
            "type": "string"
          }
        },
        "required": [
          "threshold",
          "observed",
          "window",
          "trigger_field"
        ],
        "title": "IncidentExplanationNew",
        "type": "object"
      },
      "IncidentListResponse": {
        "additionalProperties": false,
        "properties": {
          "incident_count": {
            "title": "Incident Count",
            "type": "integer"
          },
          "incidents": {
            "items": {
              "$ref": "#/components/schemas/LifecycleIncidentResponse"
            },
            "title": "Incidents",
            "type": "array"
          }
        },
        "required": [
          "incident_count",
          "incidents"
        ],
        "title": "IncidentListResponse",
        "type": "object"
      },
      "IncidentNew": {
        "properties": {
          "affected_entities": {
            "items": {
              "type": "string"
            },
            "title": "Affected Entities",
            "type": "array"
          },
          "confidence": {
            "title": "Confidence",
            "type": "number"
          },
          "created_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "evidence": {
            "$ref": "#/components/schemas/IncidentEvidenceNew"
          },
          "evidence_count": {
            "title": "Evidence Count",
            "type": "integer"
          },
          "explanation": {
            "$ref": "#/components/schemas/IncidentExplanationNew"
          },
          "first_seen": {
            "title": "First Seen",
            "type": "string"
          },
          "incident_id": {
            "title": "Incident Id",
            "type": "string"
          },
          "last_seen": {
            "title": "Last Seen",
            "type": "string"
          },
          "mitre_technique": {
            "default": "T1110",
            "title": "Mitre Technique",
            "type": "string"
          },
          "recommended_actions": {
            "items": {
              "type": "string"
            },
            "title": "Recommended Actions",
            "type": "array"
          },
          "resolution_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resolution Reason"
          },
          "severity": {
            "title": "Severity",
            "type": "string"
          },
          "source_count": {
            "title": "Source Count",
            "type": "integer"
          },
          "status": {
            "default": "open",
            "enum": [
              "open",
              "acknowledged",
              "closed"
            ],
            "title": "Status",
            "type": "string"
          },
          "subject": {
            "$ref": "#/components/schemas/IncidentSubjectNew"
          },
          "summary": {
            "title": "Summary",
            "type": "string"
          },
          "type": {
            "title": "Type",
            "type": "string"
          },
          "updated_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          }
        },
        "required": [
          "incident_id",
          "type",
          "severity",
          "confidence",
          "first_seen",
          "last_seen",
          "affected_entities",
          "evidence_count",
          "source_count",
          "summary",
          "recommended_actions",
          "explanation",
          "subject",
          "evidence"
        ],
        "title": "IncidentNew",
        "type": "object"
      },
      "IncidentPatchRequest": {
        "additionalProperties": false,
        "properties": {
          "resolution_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resolution Reason"
          },
          "status": {
            "enum": [
              "acknowledged",
              "closed"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status"
        ],
        "title": "IncidentPatchRequest",
        "type": "object"
      },
      "IncidentSubjectNew": {
        "properties": {
          "source_ip": {
            "title": "Source Ip",
            "type": "string"
          },
          "username": {
            "title": "Username",
            "type": "string"
          }
        },
        "required": [
          "source_ip",
          "username"
        ],
        "title": "IncidentSubjectNew",
        "type": "object"
      },
      "IngestResponse": {
        "additionalProperties": false,
        "properties": {
          "detection_status": {
            "enum": [
              "pending",
              "success",
              "failed"
            ],
            "title": "Detection Status",
            "type": "string"
          },
          "event_count": {
            "title": "Event Count",
            "type": "integer"
          },
          "incident_count": {
            "title": "Incident Count",
            "type": "integer"
          },
          "incidents": {
            "items": {
              "$ref": "#/components/schemas/LifecycleIncidentResponse"
            },
            "title": "Incidents",
            "type": "array"
          },
          "normalization_status": {
            "enum": [
              "pending",
              "success",
              "failed"
            ],
            "title": "Normalization Status",
            "type": "string"
          },
          "run_id": {
            "title": "Run Id",
            "type": "string"
          }
        },
        "required": [
          "run_id",
          "event_count",
          "normalization_status",
          "detection_status",
          "incident_count",
          "incidents"
        ],
        "title": "IngestResponse",
        "type": "object"
      },
      "LifecycleIncidentResponse": {
        "properties": {
          "affected_entities": {
            "items": {
              "type": "string"
            },
            "title": "Affected Entities",
            "type": "array"
          },
          "confidence": {
            "title": "Confidence",
            "type": "number"
          },
          "created_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "evidence": {
            "$ref": "#/components/schemas/IncidentEvidenceNew"
          },
          "evidence_count": {
            "title": "Evidence Count",
            "type": "integer"
          },
          "explanation": {
            "$ref": "#/components/schemas/IncidentExplanationNew"
          },
          "first_seen": {
            "title": "First Seen",
            "type": "string"
          },
          "incident_id": {
            "title": "Incident Id",
            "type": "string"
          },
          "is_stale": {
            "title": "Is Stale",
            "type": "boolean"
          },
          "last_seen": {
            "title": "Last Seen",
            "type": "string"
          },
          "mitre_technique": {
            "default": "T1110",
            "title": "Mitre Technique",
            "type": "string"
          },
          "recommended_actions": {
            "items": {
              "type": "string"
            },
            "title": "Recommended Actions",
            "type": "array"
          },
          "resolution_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resolution Reason"
          },
          "severity": {
            "title": "Severity",
            "type": "string"
          },
          "source_count": {
            "title": "Source Count",
            "type": "integer"
          },
          "status": {
            "default": "open",
            "enum": [
              "open",
              "acknowledged",
              "closed"
            ],
            "title": "Status",
            "type": "string"
          },
          "subject": {
            "$ref": "#/components/schemas/IncidentSubjectNew"
          },
          "summary": {
            "title": "Summary",
            "type": "string"
          },
          "type": {
            "title": "Type",
            "type": "string"
          },
          "updated_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          }
        },
        "required": [
          "incident_id",
          "type",
          "severity",
          "confidence",
          "first_seen",
          "last_seen",
          "affected_entities",
          "evidence_count",
          "source_count",
          "summary",
          "recommended_actions",
          "explanation",
          "subject",
          "evidence",
          "is_stale"
        ],
        "title": "LifecycleIncidentResponse",
        "type": "object"
      },
      "MetricsResponse": {
        "additionalProperties": false,
        "properties": {
          "events_by_source": {
            "additionalProperties": {
              "type": "integer"
            },
            "title": "Events By Source",
            "type": "object"
          },
          "events_ingested_total": {
            "title": "Events Ingested Total",
            "type": "integer"
          },
          "incidents_closed_total": {
            "title": "Incidents Closed Total",
            "type": "integer"
          },
          "incidents_created_total": {
            "title": "Incidents Created Total",
            "type": "integer"
          },
          "incidents_reopened_total": {
            "title": "Incidents Reopened Total",
            "type": "integer"
          },
          "incidents_total": {
            "additionalProperties": {
              "type": "integer"
            },
            "title": "Incidents Total",
            "type": "object"
          },
          "incidents_updated_total": {
            "title": "Incidents Updated Total",
            "type": "integer"
          },
          "missing_required_total": {
            "title": "Missing Required Total",
            "type": "integer"
          },
          "normalized_success_total": {
            "title": "Normalized Success Total",
            "type": "integer"
          },
          "runs_total": {
            "title": "Runs Total",
            "type": "integer"
          },
          "telemetry_rejected_total": {
            "title": "Telemetry Rejected Total",
            "type": "integer"
          }
        },
        "required": [
          "events_ingested_total",
          "normalized_success_total",
          "telemetry_rejected_total",
          "missing_required_total",
          "events_by_source",
          "incidents_total",
          "runs_total",
          "incidents_created_total",
          "incidents_updated_total",
          "incidents_reopened_total",
          "incidents_closed_total"
        ],
        "title": "MetricsResponse",
        "type": "object"
      },
      "NormalizedEventNew": {
        "properties": {
          "event_type": {
            "default": "login_attempt",
            "title": "Event Type",
            "type": "string"
          },
          "raw_source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Raw Source"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          },
          "result": {
            "title": "Result",
            "type": "string"
          },
          "source": {
            "default": "auth_service",
            "title": "Source",
            "type": "string"
          },
          "source_ip": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Ip"
          },
          "timestamp": {
            "title": "Timestamp",
            "type": "string"
          },
          "user_agent": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Agent"
          },
          "username": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Username"
          }
        },
        "required": [
          "timestamp",
          "result"
        ],
        "title": "NormalizedEventNew",
        "type": "object"
      },
      "RunsIncidentsResponse": {
        "additionalProperties": false,
        "properties": {
          "incident_count": {
            "title": "Incident Count",
            "type": "integer"
          },
          "incidents": {
            "items": {
              "$ref": "#/components/schemas/IncidentNew"
            },
            "title": "Incidents",
            "type": "array"
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          }
        },
        "required": [
          "incident_count",
          "incidents"
        ],
        "title": "RunsIncidentsResponse",
        "type": "object"
      },
      "RunsMetaResponse": {
        "additionalProperties": false,
        "properties": {
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "envelope_source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Envelope Source"
          },
          "event_count": {
            "title": "Event Count",
            "type": "integer"
          },
          "schema_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Schema Version"
          }
        },
        "required": [
          "created_at",
          "event_count"
        ],
        "title": "RunsMetaResponse",
        "type": "object"
      },
      "RunsNormalizedResponse": {
        "additionalProperties": false,
        "properties": {
          "event_count": {
            "title": "Event Count",
            "type": "integer"
          },
          "events": {
            "items": {
              "$ref": "#/components/schemas/NormalizedEventNew"
            },
            "title": "Events",
            "type": "array"
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          }
        },
        "required": [
          "event_count",
          "events"
        ],
        "title": "RunsNormalizedResponse",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Defensive Security Workflow Translator",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/health": {
      "get": {
        "operationId": "health_health_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Health"
      }
    },
    "/incidents/": {
      "get": {
        "operationId": "list_incidents_incidents__get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IncidentListResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Incidents",
        "tags": [
          "incidents"
        ]
      }
    },
    "/incidents/{incident_id}": {
      "get": {
        "operationId": "get_incident_incidents__incident_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "incident_id",
            "required": true,
            "schema": {
              "title": "Incident Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LifecycleIncidentResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Incident",
        "tags": [
          "incidents"
        ]
      },
      "patch": {
        "operationId": "patch_incident_incidents__incident_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "incident_id",
            "required": true,
            "schema": {
              "title": "Incident Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IncidentPatchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LifecycleIncidentResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Patch Incident",
        "tags": [
          "incidents"
        ]
      }
    },
    "/ingest/": {
      "post": {
        "operationId": "ingest_events_ingest__post",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IngestResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Ingest Events",
        "tags": [
          "ingest"
        ]
      }
    },
    "/metrics/": {
      "get": {
        "operationId": "get_metrics_metrics__get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MetricsResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Get Metrics",
        "tags": [
          "metrics"
        ]
      }
    },
    "/runs/": {
      "get": {
        "description": "List all available run IDs.",
        "operationId": "list_runs_runs__get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Runs",
        "tags": [
          "retrieval"
        ]
      }
    },
    "/runs/{run_id}/incidents": {
      "get": {
        "description": "Get detected incidents for a specific run.",
        "operationId": "get_incidents_runs__run_id__incidents_get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunsIncidentsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Incidents",
        "tags": [
          "retrieval"
        ]
      }
    },
    "/runs/{run_id}/meta": {
      "get": {
        "description": "Get metadata for a specific run (created_at, event_count).",
        "operationId": "get_meta_runs__run_id__meta_get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunsMetaResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Meta",
        "tags": [
          "retrieval"
        ]
      }
    },
    "/runs/{run_id}/normalized": {
      "get": {
        "description": "Get normalized events for a specific run.",
        "operationId": "get_normalized_runs__run_id__normalized_get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunsNormalizedResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Normalized",
        "tags": [
          "retrieval"
        ]
      }
    }
  }
}